Execute command Web API
Execution command API. It allows sending an arbitrary command and a list of hosts or labels where the command should be executed.
The output is returned wrapped in a JSON with a host ID where the output comes from and the type of data (stdout, stderr or teleport-error for teleport related messages).

Missing features:
* MFA support
* Saving logs to S3 - it will be added in the following PR.