dronegen: Change GHA inputs to slice from map
Change the inputs in `ghaBuildType` from a map to a slice so that it is
iterated in deterministic order. Maps are iterated in a
non-deterministic order making the command for invoking a GHA workflow
possibly change each time dronegen is run due to the input arguments
changing order.

This should eliminate spurious changes to `.drone.yaml`.