Add teleport-kube-agent-udpater OCI release worflow (#1035)
Part of: https://github.com/gravitational/teleport/issues/22732

The commit reuses the `oci-image-build` workflow to build teleport-kube-agent-updater images. It also introduces a release and a promotion workflow for updater images.

It requires a few changes on the OSS side: https://github.com/gravitational/teleport/pull/23698

- make the image buildx-compatible
- parameterize the BASE_IMAGE

I faced one issue: the kube-agent-updater image build process uses the Teleport buildbox and the version was not passed to the build command. The least awkward solution I found was:

- checking out oss teleport in the release workflow
- running a `make -f build.assets/image.mk get-buildbox-version` and capturing its output
- using `oci-image-build` image-build-args to set `BUILDBOX=${{ ... }}`

It's a bit sad to checkout the teleport oss repo twice: once in the release workflow and once in the oci-build workflow, but this avoided having to modify the oci-build-workflow too much or fork it.

Source-commit: gravitational/teleport.e@0f4e1c5444df58126c433175b1217921d180b434
Source-pull-requests: gravitational/teleport.e#1035