Fix `TestCheckAccessToRegisteredResource` flaky test
When multiple tests run in parallel and one of them replaces the module's features, it affects all the running tests.

A test was changing the modules while running in parallel and it resulted in the flakiness of `TestCheckAccessToRegisteredResource` because, since #23755, we enforce that the server has the Kubernetes feature enabled in order to upsert the server.

Fixes #24115