[v10] dronegen: Sort workflow inputs for stable output
Sort the GitHub Actions inputs when generating the `gh-trigger-workflow`
command line so that it does not randomly change order, as happens when
iterating a map directly.

This is a backport of a single commit of #24102 (drone: Switch Mac
(darwin) pipelines to GitHub Actions), as it changes dronegen, but the
pipelines themselves are not being backported to `branch/v10`.