Prevent TestService_ProxySSH_Errors flakiness
The `no access checker terminates stream` test case would sometimes
get an EOF from the server because the order of operations on the
stream is nondeterministic. Errors from getting the `authz.Context`
may occur prior to the client completing the initial send, which
result in an `io.EOF` because the server already terminated the
stream. If the client message was sent prior to the server terminating
the stream then the error from getting the `authz.Context` would
be returned. To make the test case more resilient both `io.EOF` and
`trace.AccessDenied` errors are allowed.