[v12] Disable `build-macos` and `build-windows` on PR (#24233)
This commit removes the `build-macos` and `build-windows` from the PR flow, instead delegating to the bypass job.
These jobs still run at the merge queue point.

This of course means that failures in these two jobs may not be known until the merge queue.
There is an unequestionable disadvantage in not discovering those issues until that point, but this change is being recommended because:
* Currently MacOS builds are 31% of our Teleport Actions spend (~$3,500 / week)
* Windows builds are also significant at 13% (~$1,400 / week)
* There has been relatively few failures of these jobs (without other jobs also failing)

Although merge queue verification is not ideal because it's later in the process, it is considered the most critical in ensuring that `master` remains stable.

* Make sure all bypass jobs run on `ubuntu-latest`

In a couple cases this allows the jobs to be run on a cheaper instance.