[v10] Add local guidance for Linux Server guide users (#24138)
* Add local guidance for Linux Server guide users

Fixes #12018

It might be tempting for a user to follow the Linux Server guide on a
local Docker container, only to be disappointed due to issues with
accessing Auth/Proxy service instances running locally. This guide
discourages local use of the Linux Server guide and links to other
guides the user can follow instead.

* Fix linter issues

---------

Co-authored-by: Steven Martin <steven@goteleport.com>