build: Parameterise teleport repo in mac builds (#1080)
Parameterise the teleport OSS repo in the mac release/build pipelines,
similar to what is done with the linux pipelines. This allows building
from teleport-private or any other copy of the OSS teleport repo.

This also allows the workflow to work with the dronegen pipeline for
dispatching to GitHub Actions, which requires the workflow accept the
`oss-teleport-repo` input if you want to specify a source ref to check
out.

Also parameterise the github deployment environment that contains
the secrets for signing/notarizing Mac binaries. The default is set to
`promote` which is what will be used for the workflows that are kicked
off automatically from `master`/`branch/*`. A developer can set the
parameter to `build` for a workflow run to be able to access the dev
secrets instead (`promote` secrets are not available to dev branches).

Source-commit: gravitational/teleport.e@761288c776caeea78a8e8c0adb4ea1560b73249f
Source-pull-requests: gravitational/teleport.e#1080