Synchronize Okta objects with Teleport. (#1030)
Okta groups and applications are periodically synchronized with Teleport as
types.UserGroup and types.Application objects. This implementation attempts
to weather Okta resources that aren't populated with enough information for
us to synchronize.

If applications fail to synchronize, groups will still be attempted and vice
versa. However, if there is an issue either retrieving from the Okta API or
from Teleport, the synchronization process for that resource type will stop.

Synchronization happens every 2 minutes. Right now this isn't configurable,
but could be in a future update.

Source-commit: gravitational/teleport.e@7708ab95d78d54a2c1b28ff3108850c77e0bf5f9
Source-pull-requests: gravitational/teleport.e#1030