Constrict app.FindPublicAddr client. (#24331)
The client used for `app.FindPublicAddr` only uses two calls: `GetProxies`
and `GetClusterName`. However, the client parameter interface is a
`ReadAppsAccessPoint`, which requires callers to implement far more than just
the two aforementioned functions in order to use it. This commit creates a
paired down client interface for `app.FindPublicAddr`, which only requires
the bare minimum of what's required for the function to work.