Fixes roles list/read dependency for listing desktops (#24346)
Changes in #22118 were backported incorrectly to v10 in #22344, which
caused the roles list/read dependency to be enforced for listing desktops.

These changes fix that issue by using the SessionContext's AccessChecker
to get the Roles of the current user, rather than requesting them with
the auth.ClientI which is authenticated as the user themselves.