[v11] drone: Switch Mac (darwin) pipelines to GitHub Actions
* dronegen: Sort workflow inputs for stable output

Sort the GitHub Actions inputs when generating the `gh-trigger-workflow`
command line so that it does not randomly change order, as happens when
iterating a map directly.

* dronegen: Have darwin pipelines call out to GitHub Actions

Update the darwin pipelines to run workflows on GitHub Actions instead
of locally on drone builders. This replaces four pipelines with a single
GitHub actions workflow as the one workflow builds the tarballs, Mac
packages and Mac disk images.

We continue to drive the push build from drone until we work out how
secrets are safely managed in the Teleport OSS repo.

* drone: Regenerate .drone.yml for Mac pipeline changes

To regenerate the `.drone.yml` file, first three pipelines were manually
removed:
- build-darwin-amd64-pkg
- build-darwin-amd64-pkg-tsh
- build-darwin-amd64-connect

Then
    make dronegen

was run to update the pipelines:
- push-build-darwin-amd64
- build-darwin-amd64