dronegen: Fix nesting of slack template parameter
When generating a GitHub Actions pipeline, put the slack message
template under `step.settings` instead of `step`, as the template is a
parameter to the plugin, not a configuration item of the step.

The `Template` field (now unused) has not been removed from the `step`
struct as it needs to be there to unmarshal the existing `.drone.yml`
file. Once that has been regenerated, we can remove the field.