Convert tsh ls to use cluster client (#24294)
Migrates `tsh ls` to connect to the cluster via the new
`client.ClusterClient` to help reduce latency caused by geolocation.
Doing so resulted in some tests failing due to the cluster name of
the client not being set correctly. The `ClusterClient` now only
uses the guessed client if jump hosts were provided and the inferred
cluster name is not the root cluster.