[v12] dronegen: Fix nesting of slack template parameter
* dronegen: Fix nesting of slack template parameter

When generating a GitHub Actions pipeline, put the slack message
template under `step.settings` instead of `step`, as the template is a
parameter to the plugin, not a configuration item of the step.

The `Template` field (now unused) has not been removed from the `step`
struct as it needs to be there to unmarshal the existing `.drone.yml`
file. Once that has been regenerated, we can remove the field.

* drone: Regenerate .drone.yml with correct templates

Regenerate `.drone.yml` with the correct template setting for the slack
plugin.

* dronegen: Remove unused Template field from step

Remove the unused `step.Template` field as it is no longer used. It
could not be removed earlier as it was needed to be able to unmarshal
the existing `.drone.yml` file. Now that has been done in a previous
commit, the field can be removed.

Backport: https://github.com/gravitational/teleport/pull/24370