Introduce an OktaAssignmentsGetter and use it in the watcher.
The `OktaAssignmentWatcher` has been adjusted to use an
`OktaAssignmentsGetter` interface, which is a read only interface. This is
due to the fact that the `OktaAccessPoint`, used by the Okta service, does
not fully implement the `OktaAssignments` interface (it does not implement
`DeleteAllOktaAssignments`). As the watcher only needs the read functions,
this will allow us to prevent the Okta service from having access to the
`DeleteAll` while still being able to use the `OktaAssignmentWatcher`.