fix: Don't write collected data when creating devices (#1118)
Resource-like writes, like the ones done via `tctl create`, allow users to
supply various fields that are typically system-managed. Unfortunately, they are
too lenient on collected data.

Writing collected data should be predicated on a cleared device challenge - as
the name implies, it is collected from the end-user device, thus it shouldn't be
written by other sources.

In practice we also see the following:

* For versions before v13/MDM, collected data has little useful information and
  thus can be dropped without downsides in such writes
* For versions after v13/MDM, the DeviceProfile is the field that better capture
  those semantics

Related to #826 (but not an integral part of it).

Source-commit: gravitational/teleport.e@7972de1f01f05f3513d82ee0d3183f757af623ea
Source-pull-requests: gravitational/teleport.e#1118