Fix relaxed moderator joining for Kube Access (#23674) (#24181)
This PR fixes the relaxed mode for moderator joining moderated sessions.

Previously, when the moderator didn't had access to the Kubernetes
cluster where the session he tried to join was active, Teleport would
panic because the user didn't had at least one `kubernetes_user` or
`kubernetes_group`.

This PR moves the check of kubernetes principals into the autorization
step instead of the authentication step in order to prevent the failure
of authentication followed by a panic.