Fix `TestService_ProxySSH_Errors` flakiness
The server may kill the stream prior to sending or receiving a message
so the test now checks whether an io.EOF was received in response
interacting with the server. We expect that an io.EOF should be
received either when sending or receiving a message. If both complete
successfully then we expect to receive the actual access denied error.