Write audit events before completing streams (#1132)
Write audit log entries before the very last Send on device enrollment and
authentication streams. This difference in production behavior makes tests much
simpler to reason about, as we can assume that after the last Recv happens all
server-side work is done.

The tradeoff is that there's the odd chance of recording a success but failing
the last Send afterwards. In this scenario all the server-side work is done, so
it's not a terrible tradeoff.

Fixes various flakiness issues where we observe an unexpected number of audit
events.

Source-commit: gravitational/teleport.e@e288f623c022b2e8ce6bac04fa903127ad02f2c7
Source-pull-requests: gravitational/teleport.e#1132