Add FileTransferContainer
This PR separates the absolute positioned container from the FileTransfer component. We will be adding FileTransferRequests
in the same spot, and this will allow them to be visible together without css shenanigans trying to absolutely place them together.
FileTransferRequests are separate from FileTransfers in that they interact with the tty.