helm: mount token through projected volumes when available (#20744)
* helm: mount token through projected volumes when available

Mounting ServiceAccount tokens through projected volumes instead of the
regular automout allows to reduce the token TTL from 1 year to 1 hour
when the `BoundServiceAccountTokenVolume` feature gate is on.
This helps mitigating the impact of token exfiltration (even though the
token valid 1 year still got revoked on pod termination).

Kubernetes running version 1.20 and above must
support `TokenRequest/TokenRequestProjection` (GA in 1.20). This is the
capability required to mount tokens through projected volumes. While EKS
and GKE might have supported this earlier, kubeadm, kubespray, minikube
and AKS only turned the flag on in 1.20.

Kubernetes running version 1.22 and above must support
`BoundServiceAccountTokenVolume` (beta in 1.21, GA in 1.22). This is the
capability required to issue short-lived and pod-bound tokens.

* helm: mount auth token through projected volumes when available

* fix-templates