Add cleanup time and last transition time to OktaAssignment. (#24725)
The OktaAssignment now has a cleanup time that will be specified on creation
and each action has a LastTransition time. The cleanup time specifies when
the Okta service should clean up the Okta assignment, and the last transition
time will be used to show when the action last transitioned.

It should be noted that cleanup time is maintained separately from the notion
of expiry as we want Okta assignments to stick around until they're cleaned
up, otherwise we have no record of what Teleport has done via the Okta API.

LastTransition time can be used to timeout actions that have been stuck
processing for too long.