release: Upload MacOS release artifacts to S3 (#1176)
Upload the release artifacts to the dronestaging S3 bucket before
uploading to the release server. This was thought to no longer be needed
but there are still some things that want the artifacts available from
`get.gravitational.com`, which this bucket feeds via the promotion
process.

Source-commit: gravitational/teleport.e@a48ce4d7801aa8af1d3e77b55f892b534ce6d773
Source-pull-requests: gravitational/teleport.e#1176