[v10] Backport dronegen changes for GitHub Actions (#24789)
* [v10] Make GHA pipeline structure more generic

Make the GHA pipeline structure generic instead of assuming it only
calls the linux-arm64 workflows.

This backports part of dff5cd4042c3324a39de57c214fdbe2986f54edd from
"Integrates distroless OCI publishing into drone (#22707)", but does not
backport the distroless OCI pipelines themselves.

* [v10] Removes unnecessary workflow arguments for GHA pipelines

Removes unnecessary workflow arguments from the Drone workflow
invocation. Refactors some of the `dronegen` generators to make
this easier to express insife `dronegen`

This backports "Fix OCI promotion (#22867)" without the changes
to the OCI pipelines as they have not been backported to v11.

* [v10] Increase GHA build timeout

Our ARM64 build fails quite frequently due to 30 minutes timeout. This PR increases the build timeout to one hour.

* [v10] dronegen: Sort workflow inputs for stable output

Sort the GitHub Actions inputs when generating the `gh-trigger-workflow`
command line so that it does not randomly change order, as happens when
iterating a map directly.

This is a backport of a single commit of #24102 (drone: Switch Mac
(darwin) pipelines to GitHub Actions), as it changes dronegen, but the
pipelines themselves are not being backported to `branch/v10`.

---------

Co-authored-by: Trent Clarke <trent@goteleport.com>
Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>