release: Don't error on 409 response releasing assets (#1180)
Allow a 409 error when releasing assets without erroring the whole
release. This occurs when two github workflows run in parallel, kicked
off by drone. The drone pipelines clears unpublished assets/releases at
the start of a release build allowing them to be re-uploaded, but if a
release fails (transiently) on a step and is re-run, it can be that a
github release pipeline is still running so the clearing at the start of
the drone run does not help.

i.e. the order of operations ends up being:

1. Run release pipeline (run #1)
2. Clean release/assets
3. Pipeline fails due to network glitch
4. Cancel drone release (github workflows are not cancelled!)
5. Run release pipeline (run #2)
6. Clean release/assets
7. Run #1 workflow completes successfully and uploads release/assets
8. Run #2 workflow completes successfully but fails to upload
   release/assets because the order of step 6 & 7 means they already
   exist on the release server. Release server returns a 409 and the
   workflow job fails, failing the drone pipeline.

We already handle a 409 when registering the release, but do not when
registering an asset. Refactor the release "script" to handle a 409 when
releasing assets.

While we're here, clean up:
* Don't use github expressions in the script. Pass them via env vars
  instead. This protects the script from breaking if the var contents
  might break the syntax of the script (i.e. contains a double quote)
* Don't write the release server key and cert to the filesystem -
  instead pipe it directly from the environment using bash process
  subsitution. This allows us to remove the cleanup step too.
* Don't align backslashes on contination lines - it may look neater but
  causes spurious changes (and is tedious to update) when the length of
  the lines change.

Source-commit: gravitational/teleport.e@9fb7d2be7b051329e7fc32d3edf2d2e60ffc67bd
Source-pull-requests: gravitational/teleport.e#1180