Create locks using the correct property (#24989)
* Create locks using the correct property

The "lock target" in Teleport's backend uses a different value
depending on the type of lock you want to create. For example,
to lock a node you use its UUID, but to lock a role you use its
name.

We were incorrectly always using the name to create a lock,
which appears to work fine in the UI but is not correctly
enforced on the backend.

This change adds a new required field called targetValue, allowing
us to specify the value to be used for the lock on a per-resource
basis.

Fixes gravitational/teleport-private#556

* Simplify useGetTargetData