Add headings to the Cloud FAQ (#25122)
Closes #12506

This change adds H2-level headings to the Cloud FAQ page in order to
help users navigate the page. The order of headings is based on my guess
as to the importance of each topic.

This change does **not** attempt to refresh the content of each
question/answer entry, but will make it easier to update the page in the
future since the page should be easier to navigate.