Retry search in Connect if current workspace cert has expired (#24880)
* Adjust copy of db action

* Use retryWithRelogin when getting db usernames

* useAsync: Remove unnecessary useCallback

The state setter coming from useState is always stable.

https://legacy.reactjs.org/docs/hooks-reference.html#usestate

* useAsync: Expand docs with `run` return value example

* useSearch: Rename `restrictions` to `filters`

* useSearch: Remove unnecessary useState calls

* Refactor lockOpen into pauseUserInteraction

lockOpen worked great when we were concerned only about user interaction
with a modal closing the search bar as well. However, in the next commit
I'm going to add a login modal that's shown if the search fails with a
retryable error.

In that scenario, pressing Enter in the modal wouldn't work, as it would
be captured by the window listener that ResultList adds.

To work around this problem, I refactored lockOpen into pauseUserInteraction.
It still works pretty much the same way. But then instead of having checking
isLockedOpen in the close function, we have a new addWindowEventListener
function.

addWindowEventListener automatically removes the listener after
pauseUserInteraction is called. This solves both the problem of closing
the modal and the problem of using the enter key in the modal.

* Relogin & retry resource search if current workspace cert has expired

* addWindowEventListener: Name the cleanup function

* Make addWindowEventListener a prop of ResultList