Record and verify WebAuthn RPIDs (#25238)
Record and verify the credential RPID in `WebauthnDevice`, so we can
detect and warn against accidental RPID changes.

RPID changes are not allowed by WebAuthn, so there's little Teleport can
do to mitigate them.

Users that have only "invalid" WebAuthn devices will get ["invalid
credentials"][1] errors on login. While a bit opaque, this is likely to
lead to an user reset, which is the correct fix if only a few users are
affected.

[1]:
https://github.com/gravitational/teleport/blob/7a90a0ff9943e4f536736372dbdc90d324f143a2/lib/web/apiserver.go#L2180